Convert cmd outputs yaml plan
